Usage graph bars now ng-style, UserBudgetState long

+++ ProvisioningLibrary/VolatileStorage/UserBudgetState.cs
@@ -37,7 +37,7 @@ namespace ProvisioningLibrary
                 return this.RowKey;
             }
         }
-        public double UnitsBudgetted { get; set; }
-        public double UnitsUsed { get; set; }
+        public long UnitsBudgetted { get; set; }
+        public long UnitsUsed { get; set; }
     }
 }
